requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
black==25.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
control==0.10.1
cycler==0.12.1
docutils==0.21.2
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
flit==3.12.0
flit_core==3.12.0
fonttools==4.56.0
fsspec==2025.5.1
graphviz==0.20.3
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.5
ninja==1.11.1.4
numpy==1.26.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.1
pypdf==5.4.0
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.27.1
setuptools==79.0.1
shapely==2.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
sympy==1.14.0
tifffile==2025.5.26
tomli_w==1.2.0
torch==2.7.0
torchvision==0.22.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.3
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
